requirements.txt
Flask==0.11.1
google==2.0.1
google-api-python-client==1.6.4
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-oauth==1.0.1
google-oauth2l==1.0.1
line-bot-sdk==1.5.0

